+++ packages/ui-components/src/components/LegendThreshold/LegendThresholdHorizontal.tsx
@@ -2,29 +2,16 @@ import React from "react";
 import { Group } from "@visx/group";
 import { scaleBand } from "@visx/scale";
 import { TickLabel, TickMark } from "./LegendThreshold.style";
-import { CommonLegendThresholdProps } from "./interfaces";
+import { LegendThresholdHorizontalProps } from "./interfaces";
 import { RectClipGroup } from "../RectClipGroup";
-
-export interface LegendThresholdHorizontalProps<T>
-  extends CommonLegendThresholdProps<T> {
-  /** Orientation of the bars */
-  orientation: "horizontal";
-  /** Width of the component */
-  width?: number;
-  /**
-   * Height of the colored bars. When not adding the bars, make sure that
-   * `barHeight` is set to the same value as `height`.
-   */
-  barHeight?: number;
-}
 
 /**
  * `LegendThresholdHorizontal` represents a scale with thresholds that separate
  * a set of levels. By default, the labels between each level are shown.
  */
+
 export const LegendThresholdHorizontal = <T,>({
-  height = 40,
-  barHeight = 20,
+  height = 20,
   width = 300,
   borderRadius = 10,
   items,
@@ -44,11 +31,15 @@ export const LegendThresholdHorizontal = <T,>({
   const labelTickHeight = 4;
   const tickLabelPadding = 2;
 
+  const heightWithLabels =
+    height + 20; /* An additional 20px accounts for the height of labels */
+  const totalHeight = showLabels ? heightWithLabels : height;
+
   return (
-    <svg width={width} height={height} {...otherSvgProps}>
+    <svg width={width} height={totalHeight} {...otherSvgProps}>
       <RectClipGroup
         width={width}
-        height={showLabels ? barHeight : height}
+        height={height}
         rx={borderRadius}
         ry={borderRadius}
       >
@@ -56,7 +47,7 @@ export const LegendThresholdHorizontal = <T,>({
           <rect
             key={`item-${itemIndex}`}
             x={scaleRect(itemIndex) ?? 0}
-            height={barHeight}
+            height={height}
             width={rectWidth}
             fill={getItemColor(item, itemIndex)}
           />
@@ -68,7 +59,7 @@ export const LegendThresholdHorizontal = <T,>({
           itemIndex !== items.length - 1 && (
             <Group
               key={`item-${itemIndex}`}
-              top={barHeight}
+              top={height}
               left={(scaleRect(itemIndex) ?? 0) + rectWidth}
             >
               <TickMark y1={0} y2={labelTickHeight} />
