+++ packages/ui-components/src/components/MetricLegendThreshold/MetricLegendThreshold.tsx
@@ -4,69 +4,12 @@ import { Metric } from "@actnowcoalition/metrics";
 import { assert } from "@actnowcoalition/assert";
 import { LegendThreshold } from "../LegendThreshold";
 import { useMetricCatalog } from "../MetricCatalogContext";
-
-interface LevelItem {
-  /** Level name (e.g. "High") */
-  name: string;
-  /** Level color */
-  color: string;
-  /** Description of the level */
-  description: string | undefined;
-  /** Formatted value of the threshold at the end of the current level */
-  endThreshold?: string;
-}
+import { LevelItem, MetricLegendThresholdProps } from "./interfaces";
 
 const getItemColor = (item: LevelItem) => item.color;
 const getItemLabelHorizontal = (item: LevelItem) => item.endThreshold ?? "";
 const getItemLabelVertical = (item: LevelItem) => item.name;
 const getItemSublabel = (item: LevelItem) => item.description ?? "";
-
-interface CommonMetricLegendThresholdProps {
-  /** Metric to display thresholds for. */
-  metric: Metric | string;
-  /** Whether to show level labels. Does not affect start/endLabels */
-  showLabels?: boolean;
-  /** Optional label for the left or top side of the thermometer. */
-  startLabel?: React.ReactNode;
-  /** Optional label for the right or bottom side of the thermometer. */
-  endLabel?: React.ReactNode;
-  /** Whether or not to display metric name and info. If false, only thermometer is displayed. */
-  includeOverview?: boolean;
-  /** Height of the component, including the colored bars and labels. */
-  height?: number;
-  /** Border radius of the bars */
-  borderRadius?: number;
-  /** Optional other props. */
-  otherSvgProps?: Omit<
-    React.SVGProps<SVGSVGElement>,
-    keyof MetricLegendThresholdVerticalProps
-  >;
-}
-
-export interface MetricLegendThresholdHorizontalProps
-  extends CommonMetricLegendThresholdProps {
-  /** Orientation of the bars. */
-  orientation: "horizontal";
-  /** Width of the component. */
-  width?: number;
-  /**
-   * Height of the bars. When not adding the labels, make sure that
-   * `barHeight` is set to the same value as `height`.
-   */
-  barHeight?: number;
-}
-
-export interface MetricLegendThresholdVerticalProps
-  extends CommonMetricLegendThresholdProps {
-  /** Orientation of the bars. */
-  orientation: "vertical";
-  /** Width of the bars. */
-  barWidth?: number;
-}
-
-export type MetricLegendThresholdProps =
-  | MetricLegendThresholdHorizontalProps
-  | MetricLegendThresholdVerticalProps;
 
 export const MetricLegendThreshold: React.FC<MetricLegendThresholdProps> = ({
   metric,
